% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calBiomass.R
\name{calbiomass}
\alias{calbiomass}
\title{above ground biomass calculation}
\usage{
calbiomass(data, model.Ver = "Ung2008", model.D = TRUE)
}
\arguments{
\item{data}{a data frame that includes at minimum the columns: species_code, dbh_cm, Htot_m(if using DH-model). unit of dbh_cm in centimeter and Htot_m in meter to have the estimated biomass in kg.}

\item{model.Ver}{parameter sets from 2 publications, values: "Ung2008" (default), "Lambert2005"}

\item{model.D}{logi, T: D-model(default), F: DH-model}
}
\value{
Above ground biomass (in Kg) of 4 compartments (wood, bark, branches, foliage) and tree biomass
as the sum of the 4 compartments.
}
\description{
above ground biomass calculation using Ung2008 and Lambert2005
}
\details{
For both Ung2008 and Lambert2005, We have developed D-model and DH-model depending on the availability of total height.
species_code cannot be empty, use one of the species group ("ALL", "SOFTWOOD", "HARDWOOD") for the approx. estimation.
}
\examples{
sample_data <- data.table::data.table(
  species_code = c("PICEMAR","PICEGLA"),
  treeID = c(1,2),
  dbh_cm = c(30,30),
  Htot_m = c(18,18)
)

bm <- calbiomass(data = sample_data, model.Ver = "Ung2008", model.D = TRUE)

}
\references{
Lambert, M.-C.; Ung, C.-H.; Raulier, F. 2005. Canadian national tree aboveground biomass equations. Can. J. For. Res. 35:1996-2018.

Ung, C.-H.; Bernier, P.; Guo, X.-J. 2008. Canadian national biomass equations: new parameter estimates that include British Columbia data. Can. J. For. Res 38:1123-2232.
}
\author{
Xiao Jing Guo \email{xiaojing.guo@NRCan-RNCan.gc.ca}
}
