% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list_biomass.R
\docType{data}
\name{species_list_biomass}
\alias{species_list_biomass}
\title{species list}
\format{
a data table
\describe{
\item{species_code}{species: combination of the 4-letter genus code and 3-letter species code (NFI_CODE); species group: ALL, SOFTWOOD, HARDWOOD }
\item{ENGLISH_NAME}{english name}
\item{FRENCH_NAME}{french name}
\item{SCIENTIFIC_NAME}{scientific name}
\item{MRNQ_CODE}{species code used by MRNQ}
}
}
\usage{
species_list_biomass
}
\description{
species list
}
\keyword{datasets}
